Allow early exit from selection prompts
If escape is pressed while a selection is being made, raise a
SelectionAborted exception. If Ctrl-c is pressed (raising
KeyboardInterrupt), intercept and stil raise a SelectionAborted
exception.

The raised exception holds a reference to the option that was
selected at the time the prompt was aborted. For mutli option
prompts, this will be a (possibly empty) list of choices. For a
single option or yes/no prompt, the currently highlighted option
will be included.

Fixes #12